updated Repository

+++ src/main/java/at/htl/vocabulary/controller/Repository.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 public interface Repository<T> {
     public void save(T entity);
-    public void delete(int id);
+    public void delete(Long id);
     public List<T> findAll();
-    public T findById(int id);
+    public T findById(Long id);
 
     //List<T> findByName(String name);
 }
